+++ resources/ReleaseNotes.md
@@ -5,5 +5,8 @@
 - Add processing step `mag.process.Compression` to correct for compression factor
 - Make `mag.process.Calibration` more flexible to custom variable names
 - Add property `Harness` in `mag.meta.Science` to describe sensor harness
-- Allow plotting more than one event in `mag.graphics.view.Field` and fix typos in compression event plot
+- Allow plotting more than one event in `mag.graphics.view.Field`
+- Change event type to `categorical` when data is not numeric
 - Fix issue with loading Excel files containing sensor meta data
+- Fix issue in custom events chart when event value is not `double` or `single`
+- Fix typos in compression event plot